feat: typescript migration
