+++ components/Footer.tsx
@@ -1,18 +1,18 @@
-import React from 'react';
-import Image from 'next/image';
-import styles from '../styles/styles.module.css'
+import React from "react";
+import Image from "next/image";
+import styles from "@/styles/Home.module.css";
 
 const Footer = () => {
   return (
     <div className={styles.footer}>
       <Image
-        src={"/public-assembly-white-logo.svg"}
+        src={"/pa-footer-white-logo.svg"}
         width={50}
         height={50}
         alt="public assembly logo"
       />
     </div>
   );
-}
+};
 
 export default Footer;
