+++ data/index.ts
@@ -1,0 +1,464 @@
+export const nftData = {
+  nodes: [
+    {
+      token: {
+        collectionAddress: "0xb775206e01fe2d60de85df9f18dd810998bff6d7",
+        collectionName: "Lost Echoes by Felt Zine",
+        description:
+          "Lost Echoes explores the implicit bias of anthropologists, speculative futures, and retrofuturism as researchers discover an artist’s memories in the year 3030 without context. Original art by Ina Vare with production by Felt Zine.",
+        image: {
+          url: "ipfs://QmTSorii5fA3beuQJtP5kMsWBcfWWiBjZs4JBQbz6mLuDj",
+        },
+        name: "#1 | Dream Redcurrants",
+        owner: "0x5e080d8b14c1da5936509c2c9ef0168a19304202",
+        tokenId: "1",
+        tokenUrl: "ipfs://Qmag7Hgh3C2igYajdYFtLgE132yjEgwjAda4x4HBXj8tNv/1",
+        tokenStandard: "ERC721",
+      },
+    },
+    {
+      token: {
+        collectionAddress: "0xb775206e01fe2d60de85df9f18dd810998bff6d7",
+        collectionName: "Lost Echoes by Felt Zine",
+        description:
+          "Lost Echoes explores the implicit bias of anthropologists, speculative futures, and retrofuturism as researchers discover an artist’s memories in the year 3030 without context. Original art by Ina Vare with production by Felt Zine.",
+        image: {
+          url: "ipfs://QmSvC7QXJTdsaNLppyzpkAc2Gjgt6mDBtab2zTXKdZr9v6",
+        },
+        name: "#2 | Amazing Rubber Booties",
+        owner: "0x5e080d8b14c1da5936509c2c9ef0168a19304202",
+        tokenId: "2",
+        tokenUrl: "ipfs://Qmag7Hgh3C2igYajdYFtLgE132yjEgwjAda4x4HBXj8tNv/2",
+        tokenStandard: "ERC721",
+      },
+    },
+    {
+      token: {
+        collectionAddress: "0xb775206e01fe2d60de85df9f18dd810998bff6d7",
+        collectionName: "Lost Echoes by Felt Zine",
+        description:
+          "Lost Echoes explores the implicit bias of anthropologists, speculative futures, and retrofuturism as researchers discover an artist’s memories in the year 3030 without context. Original art by Ina Vare with production by Felt Zine.",
+        image: {
+          url: "ipfs://QmNYbrQqeuA7NSQqQugvDiHummBhrJezMHLeab94my6Z71",
+        },
+        name: "#3 | Imaginary Grapes",
+        owner: "0x5e080d8b14c1da5936509c2c9ef0168a19304202",
+        tokenId: "3",
+        tokenUrl: "ipfs://Qmag7Hgh3C2igYajdYFtLgE132yjEgwjAda4x4HBXj8tNv/3",
+        tokenStandard: "ERC721",
+      },
+    },
+    {
+      token: {
+        collectionAddress: "0xb775206e01fe2d60de85df9f18dd810998bff6d7",
+        collectionName: "Lost Echoes by Felt Zine",
+        description:
+          "Lost Echoes explores the implicit bias of anthropologists, speculative futures, and retrofuturism as researchers discover an artist’s memories in the year 3030 without context. Original art by Ina Vare with production by Felt Zine.",
+        image: {
+          url: "ipfs://Qmb2DJwyP8aHW4gBp38yCYGpUpS3FDKQH4vYCEuEtS8JBF",
+        },
+        name: "#4 | Misunderstood Redcurrants",
+        owner: "0x5e080d8b14c1da5936509c2c9ef0168a19304202",
+        tokenId: "4",
+        tokenUrl: "ipfs://Qmag7Hgh3C2igYajdYFtLgE132yjEgwjAda4x4HBXj8tNv/4",
+        tokenStandard: "ERC721",
+      },
+    },
+    {
+      token: {
+        collectionAddress: "0xb775206e01fe2d60de85df9f18dd810998bff6d7",
+        collectionName: "Lost Echoes by Felt Zine",
+        description:
+          "Lost Echoes explores the implicit bias of anthropologists, speculative futures, and retrofuturism as researchers discover an artist’s memories in the year 3030 without context. Original art by Ina Vare with production by Felt Zine.",
+        image: {
+          url: "ipfs://QmccmYbBLxxQVKf5CajBQCgBGa6VVU5pfvAay6mzuCjijJ",
+        },
+        name: "#5 | Volatile Zucchini",
+        owner: "0x5e080d8b14c1da5936509c2c9ef0168a19304202",
+        tokenId: "5",
+        tokenUrl: "ipfs://Qmag7Hgh3C2igYajdYFtLgE132yjEgwjAda4x4HBXj8tNv/5",
+        tokenStandard: "ERC721",
+      },
+    },
+    {
+      token: {
+        collectionAddress: "0xb775206e01fe2d60de85df9f18dd810998bff6d7",
+        collectionName: "Lost Echoes by Felt Zine",
+        description:
+          "Lost Echoes explores the implicit bias of anthropologists, speculative futures, and retrofuturism as researchers discover an artist’s memories in the year 3030 without context. Original art by Ina Vare with production by Felt Zine.",
+        image: {
+          url: "ipfs://QmbXZ8LGfYRVqHaHwNbuBfNNoAR1wmZnJNbgvvqxWQu5a7",
+        },
+        name: "#6 | Strategic Sea Buckthorn",
+        owner: "0x5e080d8b14c1da5936509c2c9ef0168a19304202",
+        tokenId: "6",
+        tokenUrl: "ipfs://Qmag7Hgh3C2igYajdYFtLgE132yjEgwjAda4x4HBXj8tNv/6",
+        tokenStandard: "ERC721",
+      },
+    },
+    {
+      token: {
+        collectionAddress: "0xb775206e01fe2d60de85df9f18dd810998bff6d7",
+        collectionName: "Lost Echoes by Felt Zine",
+        description:
+          "Lost Echoes explores the implicit bias of anthropologists, speculative futures, and retrofuturism as researchers discover an artist’s memories in the year 3030 without context. Original art by Ina Vare with production by Felt Zine.",
+        image: {
+          url: "ipfs://QmdQrxVC7jhg63qtgfAnp39oxDmJdionhmEkinJG6X3E5v",
+        },
+        name: "#7 | Pastoral Raspberry",
+        owner: "0x5e080d8b14c1da5936509c2c9ef0168a19304202",
+        tokenId: "7",
+        tokenUrl: "ipfs://Qmag7Hgh3C2igYajdYFtLgE132yjEgwjAda4x4HBXj8tNv/7",
+        tokenStandard: "ERC721",
+      },
+    },
+    {
+      token: {
+        collectionAddress: "0xb775206e01fe2d60de85df9f18dd810998bff6d7",
+        collectionName: "Lost Echoes by Felt Zine",
+        description:
+          "Lost Echoes explores the implicit bias of anthropologists, speculative futures, and retrofuturism as researchers discover an artist’s memories in the year 3030 without context. Original art by Ina Vare with production by Felt Zine.",
+        image: {
+          url: "ipfs://QmXgnLnFU9KSjYNhA6jMAT9EADc7FkxJhbiaMnnzTjbkoi",
+        },
+        name: "#8 | Virtual Rubber Toy",
+        owner: "0x5e080d8b14c1da5936509c2c9ef0168a19304202",
+        tokenId: "8",
+        tokenUrl: "ipfs://Qmag7Hgh3C2igYajdYFtLgE132yjEgwjAda4x4HBXj8tNv/8",
+        tokenStandard: "ERC721",
+      },
+    },
+    {
+      token: {
+        collectionAddress: "0xb775206e01fe2d60de85df9f18dd810998bff6d7",
+        collectionName: "Lost Echoes by Felt Zine",
+        description:
+          "Lost Echoes explores the implicit bias of anthropologists, speculative futures, and retrofuturism as researchers discover an artist’s memories in the year 3030 without context. Original art by Ina Vare with production by Felt Zine.",
+        image: {
+          url: "ipfs://QmPiEsQpA89n6sn2sEa92nn4aQUehtm8uj2L1RYABMNxsi",
+        },
+        name: "#9 | Real Rowan Berries",
+        owner: "0x5e080d8b14c1da5936509c2c9ef0168a19304202",
+        tokenId: "9",
+        tokenUrl: "ipfs://Qmag7Hgh3C2igYajdYFtLgE132yjEgwjAda4x4HBXj8tNv/9",
+        tokenStandard: "ERC721",
+      },
+    },
+    {
+      token: {
+        collectionAddress: "0xb775206e01fe2d60de85df9f18dd810998bff6d7",
+        collectionName: "Lost Echoes by Felt Zine",
+        description:
+          "Lost Echoes explores the implicit bias of anthropologists, speculative futures, and retrofuturism as researchers discover an artist’s memories in the year 3030 without context. Original art by Ina Vare with production by Felt Zine.",
+        image: {
+          url: "ipfs://QmX27i5XsAdzoyqp6fRs5hFLzoPBpsGtj1Tj3WUg5zqdSZ",
+        },
+        name: "#10 | Gifted Rowan Berries",
+        owner: "0x5e080d8b14c1da5936509c2c9ef0168a19304202",
+        tokenId: "10",
+        tokenUrl: "ipfs://Qmag7Hgh3C2igYajdYFtLgE132yjEgwjAda4x4HBXj8tNv/10",
+        tokenStandard: "ERC721",
+      },
+    },
+  ],
+};
+
+export const market = `
+query MyQuery {
+  markets(
+    networks: {network: ETHEREUM, chain: MAINNET}
+    pagination: {limit: 10}
+    sort: {sortDirection: ASC, sortKey: NONE}
+  ) {
+    nodes {
+      market {
+        collectionAddress
+        marketType
+        marketAddress
+      }
+      token {
+        name
+        owner
+        collectionAddress
+        collectionName
+        image {
+          url
+        }
+      }
+    }
+  }
+}`;
+
+export const marketData = {
+  nodes: [
+    {
+      token: {
+        name: "Ideas – Completed",
+        owner: "0xd72cb55fd6e7d94808b27b15f6132cdfcf5a0461",
+        collectionAddress: "0x0bc2a24ce568dad89691116d5b34deb6c203f342",
+        collectionName: "Catalog",
+        image: {
+          url: "ipfs://Qmb1An14hxJSH278ENP6xSUQe9tE6VsHj7N43NM3jJNFe6",
+        },
+      },
+    },
+    {
+      token: {
+        name: "Discuss idea a lot – Completed",
+        owner: "0x0d89421d6eec0a4385f95f410732186a2ab45077",
+        collectionAddress: "0x19b703f65aa7e1e775bd06c2aa0d0d08c80f1c45",
+        collectionName: "SongADay",
+        image: {
+          url: "ipfs://bafybeibdcpcgqumxzu2k5hunijb7fbcrzvpzmsbghhuo4orqq4gy2b65d4",
+        },
+      },
+    },
+    {
+      token: {
+        name: "Design tool – Completed",
+        owner: "0xe468ce99444174bd3bbbed09209577d25d1ad673",
+        collectionAddress: "0xabefbc9fd2f806065b4f3c237d4b59d9a97bcac7",
+        collectionName: "Zora",
+        image: {
+          url: null,
+        },
+      },
+    },
+    {
+      token: {
+        name: "Design tool – Completed",
+        owner: "0x5090c4fead5be112b643bc75d61bf42339675448",
+        collectionAddress: "0x19b703f65aa7e1e775bd06c2aa0d0d08c80f1c45",
+        collectionName: "SongADay",
+        image: {
+          url: "ipfs://bafybeidudwizwswovbthifw2nmythsky5on2ctepm4zotzfzj25wq7nofy",
+        },
+      },
+    },
+    {
+      token: {
+        name: "Write V1 Protocol – Completed",
+        owner: "0xe06f4f1c019551e6ca350f56b5d1300dfc24b5f0",
+        collectionAddress: "0x19b703f65aa7e1e775bd06c2aa0d0d08c80f1c45",
+        collectionName: "SongADay",
+        image: {
+          url: "ipfs://bafybeigiyvgbin7iwo7xueoplmoagpoxsc7kj2lpjukh5wnspjqqrujoua",
+        },
+      },
+    },
+    {
+      token: {
+        name: "Build tool – Completed",
+        owner: "0xb2e5751f5cc32a194fff28d247288d1aa90a8fb5",
+        collectionAddress: "0xff9c1b15b16263c61d017ee9f65c50e4ae0113d7",
+        collectionName: "Loot",
+        image: {
+          url: "data:image/svg+xml;base64,PHN2ZyB4bWxucz0iaHR0cDovL3d3dy53My5vcmcvMjAwMC9zdmciIHByZXNlcnZlQXNwZWN0UmF0aW89InhNaW5ZTWluIG1lZXQiIHZpZXdCb3g9IjAgMCAzNTAgMzUwIj48c3R5bGU+LmJhc2UgeyBmaWxsOiB3aGl0ZTsgZm9udC1mYW1pbHk6IHNlcmlmOyBmb250LXNpemU6IDE0cHg7IH08L3N0eWxlPjxyZWN0IHdpZHRoPSIxMDAlIiBoZWlnaHQ9IjEwMCUiIGZpbGw9ImJsYWNrIiAvPjx0ZXh0IHg9IjEwIiB5PSIyMCIgY2xhc3M9ImJhc2UiPlRvbWUgb2YgdGhlIEZveDwvdGV4dD48dGV4dCB4PSIxMCIgeT0iNDAiIGNsYXNzPSJiYXNlIj5TaGlydCBvZiB0aGUgRm94PC90ZXh0Pjx0ZXh0IHg9IjEwIiB5PSI2MCIgY2xhc3M9ImJhc2UiPk9ybmF0ZSBIZWxtIG9mIHRoZSBGb3g8L3RleHQ+PHRleHQgeD0iMTAiIHk9IjgwIiBjbGFzcz0iYmFzZSI+TGluZW4gU2FzaDwvdGV4dD48dGV4dCB4PSIxMCIgeT0iMTAwIiBjbGFzcz0iYmFzZSI+TGluZW4gU2hvZXM8L3RleHQ+PHRleHQgeD0iMTAiIHk9IjEyMCIgY2xhc3M9ImJhc2UiPiJHcmltIE1vb24iIERlbW9uJ3MgSGFuZHMgb2YgUHJvdGVjdGlvbiArMTwvdGV4dD48dGV4dCB4PSIxMCIgeT0iMTQwIiBjbGFzcz0iYmFzZSI+IkRlYXRoIFdoaXNwZXIiIEFtdWxldCBvZiB0aGUgRm94PC90ZXh0Pjx0ZXh0IHg9IjEwIiB5PSIxNjAiIGNsYXNzPSJiYXNlIj5UaXRhbml1bSBSaW5nPC90ZXh0Pjwvc3ZnPg==",
+        },
+      },
+    },
+    {
+      token: {
+        name: "Continue to discuss a lot – Completed",
+        owner: "0x3ddea6385b7d0b75a6dae639bc39cb315fcc0f74",
+        collectionAddress: "0xabefbc9fd2f806065b4f3c237d4b59d9a97bcac7",
+        collectionName: "Zora",
+        image: {
+          url: "https://zora-prod.mypinata.cloud/ipfs/bafybeidxg53w4o4br6gqmhbsdsqw4q5y67un4x3pdrnpzi3gigq2ezpni4",
+        },
+      },
+    },
+    {
+      token: {
+        name: "Submit solidity for audit – Completed",
+        owner: "0xb9354777fd3019df43fa36a85e76ee3fe2f8bc35",
+        collectionAddress: "0xabefbc9fd2f806065b4f3c237d4b59d9a97bcac7",
+        collectionName: "Zora",
+        image: {
+          url: null,
+        },
+      },
+    },
+    {
+      token: {
+        name: "Finalize audit – Completed",
+        owner: "0xe468ce99444174bd3bbbed09209577d25d1ad673",
+        collectionAddress: "0xabefbc9fd2f806065b4f3c237d4b59d9a97bcac7",
+        collectionName: "Zora",
+        image: {
+          url: null,
+        },
+      },
+    },
+    {
+      token: {
+        name: "Create Builder DAO art – Completed",
+        owner: "0xe468ce99444174bd3bbbed09209577d25d1ad673",
+        collectionAddress: "0xabefbc9fd2f806065b4f3c237d4b59d9a97bcac7",
+        collectionName: "Zora",
+        image: {
+          url: null,
+        },
+      },
+    },
+    {
+      token: {
+        name: "Deploy on Mainnet – Completed",
+        owner: "0xe468ce99444174bd3bbbed09209577d25d1ad673",
+        collectionAddress: "0xabefbc9fd2f806065b4f3c237d4b59d9a97bcac7",
+        collectionName: "Zora",
+        image: {
+          url: null,
+        },
+      },
+    },
+    {
+      token: {
+        name: "Deploy to production – Completed",
+        owner: "0xe468ce99444174bd3bbbed09209577d25d1ad673",
+        collectionAddress: "0xabefbc9fd2f806065b4f3c237d4b59d9a97bcac7",
+        collectionName: "Zora",
+        image: {
+          url: null,
+        },
+      },
+    },
+    {
+      token: {
+        name: "Aquire nounsbuilder.eth – Completed",
+        owner: "0xe468ce99444174bd3bbbed09209577d25d1ad673",
+        collectionAddress: "0xabefbc9fd2f806065b4f3c237d4b59d9a97bcac7",
+        collectionName: "Zora",
+        image: {
+          url: null,
+        },
+      },
+    },
+    {
+      token: {
+        name: "20 Nouns in zora.eth – Completed",
+        owner: "0xe468ce99444174bd3bbbed09209577d25d1ad673",
+        collectionAddress: "0xabefbc9fd2f806065b4f3c237d4b59d9a97bcac7",
+        collectionName: "Zora",
+        image: {
+          url: null,
+        },
+      },
+    },
+    {
+      token: {
+        name: "Sumit proposal onchain – 1 March",
+        owner: "0xe468ce99444174bd3bbbed09209577d25d1ad673",
+        collectionAddress: "0xabefbc9fd2f806065b4f3c237d4b59d9a97bcac7",
+        collectionName: "Zora",
+        image: {
+          url: null,
+        },
+      },
+    },
+    {
+      token: {
+        name: "Proposal excecuted – ~ 7 March",
+        owner: "0xe468ce99444174bd3bbbed09209577d25d1ad673",
+        collectionAddress: "0xabefbc9fd2f806065b4f3c237d4b59d9a97bcac7",
+        collectionName: "Zora",
+        image: {
+          url: null,
+        },
+      },
+    },
+    {
+      token: {
+        name: "Buidler DAO starts – ~15 March",
+        owner: "0xe468ce99444174bd3bbbed09209577d25d1ad673",
+        collectionAddress: "0xabefbc9fd2f806065b4f3c237d4b59d9a97bcac7",
+        collectionName: "Zora",
+        image: {
+          url: null,
+        },
+      },
+    },
+    {
+      token: {
+        name: "??? – Imagine",
+        owner: "0xe468ce99444174bd3bbbed09209577d25d1ad673",
+        collectionAddress: "0xabefbc9fd2f806065b4f3c237d4b59d9a97bcac7",
+        collectionName: "Zora",
+        image: {
+          url: null,
+        },
+      },
+    },
+  ],
+};
+
+export const ideas = [
+  {
+    name: "Ideas – Completed",
+    url: "",
+  },
+  {
+    name: "Discuss idea a lot – Completed",
+    url: "",
+  },
+  {
+    name: "Design tool – Completed",
+    url: "",
+  },
+  {
+    name: "Write V1 Protocol – Completed",
+    url: "",
+  },
+  {
+    name: "Build tool – Completed",
+    url: "",
+  },
+  {
+    name: "Continue to discuss a lot – Completed",
+    url: "",
+  },
+  {
+    name: "Submit solidity for audit – Completed",
+    url: "",
+  },
+  {
+    name: "Finalize audit – Completed",
+    url: "",
+  },
+  {
+    name: "Create Builder DAO art – Completed",
+    url: "",
+  },
+  {
+    name: "Deploy on Mainnet – Completed",
+    url: "",
+  },
+  {
+    name: "Deploy to production – Completed",
+    url: "",
+  },
+  {
+    name: "Aquire nounsbuilder.eth – Completed",
+    url: "",
+  },
+  {
+    name: "20 Nouns in zora.eth – Completed",
+    url: "",
+  },
+  {
+    name: "Sumit proposal onchain – 1 March",
+    url: "",
+  },
+  {
+    name: "Proposal excecuted – ~ 7 March",
+    url: "",
+  },
+  {
+    name: "Buidler DAO starts – ~15 March",
+    url: "",
+  },
+  {
+    name: "??? – Imagine",
+    url: "",
+  },
+];
