--- lib/data.js
@@ -1,165 +1,0 @@
-export const nftData = {
-  nodes: [
-    {
-      token: {
-        collectionAddress: "0xb775206e01fe2d60de85df9f18dd810998bff6d7",
-        collectionName: "Lost Echoes by Felt Zine",
-        description:
-          "Lost Echoes explores the implicit bias of anthropologists, speculative futures, and retrofuturism as researchers discover an artist’s memories in the year 3030 without context. Original art by Ina Vare with production by Felt Zine.",
-        image: {
-          url: "ipfs://QmTSorii5fA3beuQJtP5kMsWBcfWWiBjZs4JBQbz6mLuDj",
-        },
-        name: "#1 | Dream Redcurrants",
-        owner: "0x5e080d8b14c1da5936509c2c9ef0168a19304202",
-        tokenId: "1",
-        tokenUrl: "ipfs://Qmag7Hgh3C2igYajdYFtLgE132yjEgwjAda4x4HBXj8tNv/1",
-        tokenStandard: "ERC721",
-      },
-    },
-    {
-      token: {
-        collectionAddress: "0xb775206e01fe2d60de85df9f18dd810998bff6d7",
-        collectionName: "Lost Echoes by Felt Zine",
-        description:
-          "Lost Echoes explores the implicit bias of anthropologists, speculative futures, and retrofuturism as researchers discover an artist’s memories in the year 3030 without context. Original art by Ina Vare with production by Felt Zine.",
-        image: {
-          url: "ipfs://QmSvC7QXJTdsaNLppyzpkAc2Gjgt6mDBtab2zTXKdZr9v6",
-        },
-        name: "#2 | Amazing Rubber Booties",
-        owner: "0x5e080d8b14c1da5936509c2c9ef0168a19304202",
-        tokenId: "2",
-        tokenUrl: "ipfs://Qmag7Hgh3C2igYajdYFtLgE132yjEgwjAda4x4HBXj8tNv/2",
-        tokenStandard: "ERC721",
-      },
-    },
-    {
-      token: {
-        collectionAddress: "0xb775206e01fe2d60de85df9f18dd810998bff6d7",
-        collectionName: "Lost Echoes by Felt Zine",
-        description:
-          "Lost Echoes explores the implicit bias of anthropologists, speculative futures, and retrofuturism as researchers discover an artist’s memories in the year 3030 without context. Original art by Ina Vare with production by Felt Zine.",
-        image: {
-          url: "ipfs://QmNYbrQqeuA7NSQqQugvDiHummBhrJezMHLeab94my6Z71",
-        },
-        name: "#3 | Imaginary Grapes",
-        owner: "0x5e080d8b14c1da5936509c2c9ef0168a19304202",
-        tokenId: "3",
-        tokenUrl: "ipfs://Qmag7Hgh3C2igYajdYFtLgE132yjEgwjAda4x4HBXj8tNv/3",
-        tokenStandard: "ERC721",
-      },
-    },
-    {
-      token: {
-        collectionAddress: "0xb775206e01fe2d60de85df9f18dd810998bff6d7",
-        collectionName: "Lost Echoes by Felt Zine",
-        description:
-          "Lost Echoes explores the implicit bias of anthropologists, speculative futures, and retrofuturism as researchers discover an artist’s memories in the year 3030 without context. Original art by Ina Vare with production by Felt Zine.",
-        image: {
-          url: "ipfs://Qmb2DJwyP8aHW4gBp38yCYGpUpS3FDKQH4vYCEuEtS8JBF",
-        },
-        name: "#4 | Misunderstood Redcurrants",
-        owner: "0x5e080d8b14c1da5936509c2c9ef0168a19304202",
-        tokenId: "4",
-        tokenUrl: "ipfs://Qmag7Hgh3C2igYajdYFtLgE132yjEgwjAda4x4HBXj8tNv/4",
-        tokenStandard: "ERC721",
-      },
-    },
-    {
-      token: {
-        collectionAddress: "0xb775206e01fe2d60de85df9f18dd810998bff6d7",
-        collectionName: "Lost Echoes by Felt Zine",
-        description:
-          "Lost Echoes explores the implicit bias of anthropologists, speculative futures, and retrofuturism as researchers discover an artist’s memories in the year 3030 without context. Original art by Ina Vare with production by Felt Zine.",
-        image: {
-          url: "ipfs://QmccmYbBLxxQVKf5CajBQCgBGa6VVU5pfvAay6mzuCjijJ",
-        },
-        name: "#5 | Volatile Zucchini",
-        owner: "0x5e080d8b14c1da5936509c2c9ef0168a19304202",
-        tokenId: "5",
-        tokenUrl: "ipfs://Qmag7Hgh3C2igYajdYFtLgE132yjEgwjAda4x4HBXj8tNv/5",
-        tokenStandard: "ERC721",
-      },
-    },
-    {
-      token: {
-        collectionAddress: "0xb775206e01fe2d60de85df9f18dd810998bff6d7",
-        collectionName: "Lost Echoes by Felt Zine",
-        description:
-          "Lost Echoes explores the implicit bias of anthropologists, speculative futures, and retrofuturism as researchers discover an artist’s memories in the year 3030 without context. Original art by Ina Vare with production by Felt Zine.",
-        image: {
-          url: "ipfs://QmbXZ8LGfYRVqHaHwNbuBfNNoAR1wmZnJNbgvvqxWQu5a7",
-        },
-        name: "#6 | Strategic Sea Buckthorn",
-        owner: "0x5e080d8b14c1da5936509c2c9ef0168a19304202",
-        tokenId: "6",
-        tokenUrl: "ipfs://Qmag7Hgh3C2igYajdYFtLgE132yjEgwjAda4x4HBXj8tNv/6",
-        tokenStandard: "ERC721",
-      },
-    },
-    {
-      token: {
-        collectionAddress: "0xb775206e01fe2d60de85df9f18dd810998bff6d7",
-        collectionName: "Lost Echoes by Felt Zine",
-        description:
-          "Lost Echoes explores the implicit bias of anthropologists, speculative futures, and retrofuturism as researchers discover an artist’s memories in the year 3030 without context. Original art by Ina Vare with production by Felt Zine.",
-        image: {
-          url: "ipfs://QmdQrxVC7jhg63qtgfAnp39oxDmJdionhmEkinJG6X3E5v",
-        },
-        name: "#7 | Pastoral Raspberry",
-        owner: "0x5e080d8b14c1da5936509c2c9ef0168a19304202",
-        tokenId: "7",
-        tokenUrl: "ipfs://Qmag7Hgh3C2igYajdYFtLgE132yjEgwjAda4x4HBXj8tNv/7",
-        tokenStandard: "ERC721",
-      },
-    },
-    {
-      token: {
-        collectionAddress: "0xb775206e01fe2d60de85df9f18dd810998bff6d7",
-        collectionName: "Lost Echoes by Felt Zine",
-        description:
-          "Lost Echoes explores the implicit bias of anthropologists, speculative futures, and retrofuturism as researchers discover an artist’s memories in the year 3030 without context. Original art by Ina Vare with production by Felt Zine.",
-        image: {
-          url: "ipfs://QmXgnLnFU9KSjYNhA6jMAT9EADc7FkxJhbiaMnnzTjbkoi",
-        },
-        name: "#8 | Virtual Rubber Toy",
-        owner: "0x5e080d8b14c1da5936509c2c9ef0168a19304202",
-        tokenId: "8",
-        tokenUrl: "ipfs://Qmag7Hgh3C2igYajdYFtLgE132yjEgwjAda4x4HBXj8tNv/8",
-        tokenStandard: "ERC721",
-      },
-    },
-    {
-      token: {
-        collectionAddress: "0xb775206e01fe2d60de85df9f18dd810998bff6d7",
-        collectionName: "Lost Echoes by Felt Zine",
-        description:
-          "Lost Echoes explores the implicit bias of anthropologists, speculative futures, and retrofuturism as researchers discover an artist’s memories in the year 3030 without context. Original art by Ina Vare with production by Felt Zine.",
-        image: {
-          url: "ipfs://QmPiEsQpA89n6sn2sEa92nn4aQUehtm8uj2L1RYABMNxsi",
-        },
-        name: "#9 | Real Rowan Berries",
-        owner: "0x5e080d8b14c1da5936509c2c9ef0168a19304202",
-        tokenId: "9",
-        tokenUrl: "ipfs://Qmag7Hgh3C2igYajdYFtLgE132yjEgwjAda4x4HBXj8tNv/9",
-        tokenStandard: "ERC721",
-      },
-    },
-    {
-      token: {
-        collectionAddress: "0xb775206e01fe2d60de85df9f18dd810998bff6d7",
-        collectionName: "Lost Echoes by Felt Zine",
-        description:
-          "Lost Echoes explores the implicit bias of anthropologists, speculative futures, and retrofuturism as researchers discover an artist’s memories in the year 3030 without context. Original art by Ina Vare with production by Felt Zine.",
-        image: {
-          url: "ipfs://QmX27i5XsAdzoyqp6fRs5hFLzoPBpsGtj1Tj3WUg5zqdSZ",
-        },
-        name: "#10 | Gifted Rowan Berries",
-        owner: "0x5e080d8b14c1da5936509c2c9ef0168a19304202",
-        tokenId: "10",
-        tokenUrl:
-          "ipfs://Qmag7Hgh3C2igYajdYFtLgE132yjEgwjAda4x4HBXj8tNv/10",
-        tokenStandard: "ERC721",
-      },
-    }
-  ],
-}
