+++ next.config.js
@@ -1,8 +1,6 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-  experimental: {
-    appDir: true,
-  }
+  reactStrictMode: true,
 }
 
 module.exports = nextConfig
