+++ package.json
@@ -1,4 +1,7 @@
 {
+  "name": "pa-mint-house",
+  "version": "0.1.0",
+  "private": true,
   "scripts": {
     "dev": "next dev",
     "build": "next build",
@@ -6,15 +9,17 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "ethers": "^6.0.2",
-    "graphql": "^16.6.0",
-    "next": "^13.1.6",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "urql": "^3.0.3"
-  },
-  "devDependencies": {
+    "@next/font": "13.1.6",
+    "@types/node": "18.13.0",
+    "@types/react": "18.0.28",
+    "@types/react-dom": "18.0.10",
     "eslint": "8.34.0",
-    "eslint-config-next": "^13.1.6"
+    "eslint-config-next": "13.1.6",
+    "graphql": "^16.6.0",
+    "next": "13.1.6",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "typescript": "4.9.5",
+    "urql": "^3.0.3"
   }
 }
