--- pages/_app.js
@@ -1,6 +1,0 @@
-import React from 'react';
-import '../styles/global.css'
-
-export default function App({Component, pageProps}) {
-  return <Component {...pageProps} />;
-};
