--- styles/global.css
@@ -1,12 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=IBM+Plex+Mono&display=swap');
-@import url('https://fonts.googleapis.com/css2?family=Poppins:wght@900&display=swap');
-@import url('https://fonts.googleapis.com/css2?family=Space+Mono&display=swap');
-
-* {
-  font-family: 'Space Mono', sans-serif;
-  color: #1E1F22;
-}
-
-audio {
-  display: block;
-}
