--- styles/grid.module.css
@@ -1,54 +1,0 @@
-.detailsPage {
-  /* background: url("/public-ants-bw.png"); */
-  background-size: 1000px 1000px;
-  height: 97vh;
-}
-
-.grid > div {
-  /* background: #e5e6ea; */
-  font-size: 32px;
-  border-radius: 5px;
-  text-align: center;
-  overflow: hidden;
-}
-
-.grid > div:nth-child(even) {
-  /* background: #adb1b8; */
-}
-
-.grid {
-  display: grid;
-  grid-auto-rows: 70px;
-  grid-auto-columns: 20px 100px;
-  grid-gap: 10px;
-  justify-content: center;
-}
-
-.one {
-  grid-column: 1/10;
-  grid-row: 1/2;
-}
-
-.three {
-  grid-column: 1/10;
-  grid-row: 2/7;
-  background-size: cover;
-}
-
-.four {
-  grid-column: 1/3;
-}
-
-.six {
-  grid-column: 3/6;
-}
-
-.cardImage {
-  display: flex;
-  background-size: cover;
-}
-
-.seven {
-  grid-column: 6/10;
-  grid-row: 7/8;
-}
